+++ src/main/java/goorm_runner/backend/member/application/MemberService.java
@@ -3,6 +3,7 @@ package goorm_runner.backend.member.application;
 import goorm_runner.backend.global.ErrorCode;
 import goorm_runner.backend.member.application.exception.MemberException;
 import goorm_runner.backend.member.domain.Member;
+import goorm_runner.backend.member.domain.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
