+++ src/main/java/goorm_runner/backend/member/domain/AuthorityRepository.java
@@ -1,7 +1,5 @@
-package goorm_runner.backend.member.application;
+package goorm_runner.backend.member.domain;
 
-import goorm_runner.backend.member.domain.Authority;
-import goorm_runner.backend.member.domain.AuthorityType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
