+++ src/main/java/goorm_runner/backend/member/domain/MemberAuthorityRepository.java
@@ -1,6 +1,5 @@
-package goorm_runner.backend.member.application;
+package goorm_runner.backend.member.domain;
 
-import goorm_runner.backend.member.domain.MemberAuthority;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MemberAuthorityRepository extends JpaRepository<MemberAuthority, Long> {
