+++ src/main/java/goorm_runner/backend/post/application/PostService.java
@@ -1,8 +1,8 @@
 package goorm_runner.backend.post.application;
 
-import goorm_runner.backend.member.application.MemberRepository;
 import goorm_runner.backend.member.application.exception.MemberException;
 import goorm_runner.backend.member.domain.Member;
+import goorm_runner.backend.member.domain.MemberRepository;
 import goorm_runner.backend.post.application.exception.PostException;
 import goorm_runner.backend.post.domain.Category;
 import goorm_runner.backend.post.domain.Post;
